test #140
@billmeek let me know if you still run into the same issue with .27. If
so, dump your xml from http://localhost:6544/Dvr/GetRecordScheduleList
and post it here.  I will need to diff it from mine on .26 to see what
fields might be different.

+++ src/org/mythtv/client/ui/dvr/RecordingRuleActivity.java
@@ -38,7 +38,7 @@ public class RecordingRuleActivity extends AbstractMythtvFragmentActivity {
 
 	private RecordingRuleFragment recordingRuleFragment = null;
 
-	private Integer recordingRuleId;
+	private Long recordingRuleId;
 	
 	// ***************************************
 	// Activity methods
@@ -55,7 +55,7 @@ public class RecordingRuleActivity extends AbstractMythtvFragmentActivity {
 		super.onCreate( savedInstanceState );
 
 		Bundle extras = getIntent().getExtras(); 
-		recordingRuleId = extras.getInt( EXTRA_RECORDING_RULE_KEY );
+		recordingRuleId = extras.getLong( EXTRA_RECORDING_RULE_KEY );
 		
 		setContentView( R.layout.fragment_dvr_recording_rule );
